fix testRequest, some of the Test Concept was missing due non existence of reference for category concept
