+++ api/src/main/java/org/openmrs/module/poc/api/common/util/OPENMRSUUIDs.java
@@ -42,4 +42,8 @@ public interface OPENMRSUUIDs {
 	String MISAU_LABORATORIO_ENCOUNTER_TYPE_UUID = "e2790f68-1d5f-11e0-b929-000c29ad1d07";
 	
 	String LABORATORIO_GERAL_FORM_UUID = "8377e4ff-d0fe-44a5-81c3-74c9040fd5f8";
+	
+	String RAPID_PLASMA_REAGIN_CONCEP_UUID = "e1dd332c-1d5f-11e0-b929-000c29ad1d07";
+	
+	String DVRL_CONCEPT_UUID = "e1d14ce2-1d5f-11e0-b929-000c29ad1d07";
 }
